modularization part 2 + comments
